Add simple tests and small refactors
